Added CMakeLists script for libjasper.

+++ oss/jasper/src/libjasper/jp2/jp2_cod.c
@@ -490,7 +490,7 @@ int jp2_box_put(jp2_box_t *box, jas_stream_t *out)
 		box->len = jas_stream_tell(tmpstream) + JP2_BOX_HDRLEN(false);
 		jas_stream_rewind(tmpstream);
 	}
-	extlen = (box->len >= (((uint_fast64_t)1) << 32)) != 0;
+	extlen = ((uint_fast64_t)box->len >= (((uint_fast64_t)1) << 32)) != 0;
 	if (jp2_putuint32(out, extlen ? 1 : box->len)) {
 		goto error;
 	}
